backend/requirements.txt
dbt-snowflake==1.7.0
snowflake-connector-python
pandas
pyarrow
apache-airflow
openpyxl